Move Key Connector convert to own command, set usesKeyConnector after conversion
